requirements.txt
openai
langchain
langchain_community
langchain_openai
langchain_core
streamlit
python-dotenv
langserve
fastapi
uvicorn
sse_starlette
langchain-ollama
bs4
pypdf
chromadb
faiss-cpu
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
groq